+++ Readers/BamReader.cs
@@ -18,9 +18,9 @@ namespace Ovation.FasterQC.Net
 
         private bool disposedValue;
 
-        private int sequencesRead = 0;
+        private ulong sequencesRead = 0;
 
-        public int SequencesRead => sequencesRead;
+        public ulong SequencesRead => sequencesRead;
 
         public BamReader(string bam)
         {
