+++ Readers/SamReader.cs
@@ -19,9 +19,9 @@ namespace Ovation.FasterQC.Net
 
         private bool disposedValue;
 
-        private int sequencesRead = 0;
+        private ulong sequencesRead = 0;
 
-        public int SequencesRead => sequencesRead;
+        public ulong SequencesRead => sequencesRead;
 
         public SamReader(string sam, bool gzipped = true)
         {
